# api/src/static/api/api.yml
openapi: 3.0.0
info:
  title: User Service Database API
  description: 'Api rest to register users in a database for applications. Go to [Home page](http://localhost:5000/) for more information.'
  version: 1.0.0
  termsOfService: 'http://localhost:5000'
  contact:
    email: rodrjuniorsantos@gmail.com
servers:
  - url: 'http://localhost:5000'
    description: 'Primary server, e.g. Main (production) server'
components:
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: api-token
paths:
...
  /v1/api/entities/:
    post:
      summary: Create Profession Function.
      security:
        - APIKeyHeader: []
      tags:
        - Entity
      requestBody:
        description: Entity Functions
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - entity
              properties:
                entity:
                  type: object
                  properties:
                    name:
                      type: string
                    document:
                      type: string
                    deleted_app:
                      type: integer
                    created_app:
                      type: integer
                    modified_app:
                      type: integer
                image:
                  type: array
                  items:
                    type: string
                    format: binary   
                   
      responses:
        '200':
          description: Entity successfully registered
        '400':
          description: Error in request
        '401':
          description: Error in request image
        '402':
          description: Missing data
        '403':
          description: Document already exists for this user, please supply another document
        '404':
          description: There is already a registration for this user
    get:
      summary: Get all entities Function
      security:
        - APIKeyHeader: []
      tags:
        - Entity
      
      responses:
        '200':
          description: Returns all entities
        '400':
          description: User not found
        '401':
          description: Permission denied
  /v1/api/entities/{entity_id}:
    get:
      summary: Gets a entity by ID.
      security:
        - APIKeyHeader: []
      tags:
        - Entity
      parameters:
        - in: path
          name: entity_id
          required: true
          schema:
            type: integer
            minimum: 1
            description: The entity ID.

      responses:
        '200':
          description: Data this entity successfully
        '400':
          description: user not found
        '401':
          description: Permission denied
    put:
      summary: Update A Entity.
      security:
        - APIKeyHeader: []
      tags:
        - Entity
      parameters:
        - in: path
          name: entity_id
          required: true
          schema:
            type: integer
            minimum: 1
            description: The Entity ID.
      requestBody:
        description: Entity Functions
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - entity
              properties:
                entity:
                  type: object
                  properties:
                    name:
                      type: string
                    document:
                      type: string
                    deleted_app:
                      type: integer
                    created_app:
                      type: integer
                    modified_app:
                      type: integer
                image:
                  type: array
                  items:
                    type: string
                    format: binary  
      responses:
        '200':
          description: Entity successfully update
        '400':
          description: Error in request
        '401':
          description: Error in request image
        '402':
          description: Entity not found
        '403':
          description: Permission denied
        '404':
          description: Missing data
        '405':
          description: Document already exists for this user, please supply another document
        '406':
          description: There is already a registration for this user
    delete:
      summary: Delete a entity by ID.
      security:
        - APIKeyHeader: []
      tags:
        - Entity
      parameters:
        - in: path
          name: entity_id
          required: true
          schema:
            type: integer
            minimum: 1
            description: The entity ID.
      responses:
        '200':
          description: Entity successfully deleted
        '400':
          description: Entity not found
        '401':
          description: Permission denied
...